Create Calculator.cpp
The program should prompt the user to make a selection from the menu which operation is to be done. After the user has made their
selection the program should ask for the input values; note that some operations only need one input
value, whilst others need two. After the user enters the input values, the program should perform the
calculation and output the result. The program should then loop back and again prompt the user to make
a selection. The program should continue this process until the user quits.