Create fillArrayDynamicMemory.cpp
Write a function called RandomArrayFill that inputs (i.e., takes a parameter) an integer array, and
also that inputs (i.e., takes a parameter) an integer, which contains the size of the array. The function is
then to write a random number, in the range [0, 100], to each element of the array.
Do the same this time using a vector.