fractionClass
Define and implement a Fraction class with the following specifications:
•Contains a floating-point data member representing the numerator and contains a floating-point
data member representing the denominator.
•Contains a default constructor that takes no parameters and which initializes the fraction to 0 1.
•Contains a non-default constructor which takes a floating-point parameter that specifies the
numerator, and takes another floating-point point parameter that specifies the denominator.
•Overloads the arithmetic operators (+, -, *, /) to perform fraction addition, fraction subtraction,
fraction multiplication, and fraction division. (You do not need to reduce the fraction, but you
can if you are motivated to do so.)
•Overloads the relational operators (==, !=, <, >, <=, >=) so that, for any two Fractions A and B,
we can determine if A == B, A != B, A < B, A > B, A <= B, or A >= B.